requirement.txt
python-nmap
requests
